Split classes into separate file to improve readability
