+++ urchin/__init__.py
@@ -1,12 +1,10 @@
-from .urdf import (URDFType, URDFTypeWithMesh,
-                   Box, Cylinder, Sphere, Mesh, Geometry,
-                   Texture, Material,
-                   Collision, Visual, Inertial,
-                   JointCalibration, JointDynamics, JointLimit, JointMimic,
-                   SafetyController, Actuator, TransmissionJoint,
-                   Transmission, Joint, Link, URDF)
-from .utils import (rpy_to_matrix, matrix_to_rpy, xyz_rpy_to_matrix,
-                    matrix_to_xyz_rpy)
+from .base import URDFType, URDFTypeWithMesh
+from .joint import JointCalibration, JointDynamics, JointLimit, JointMimic, SafetyController, Joint
+from .transmission import Actuator, TransmissionJoint, Transmission
+from .link import Box, Cylinder, Sphere, Mesh, Geometry, Collision, Visual, Inertial, Link
+from .material import Texture, Material
+from .urdf import URDF
+from .utils import (rpy_to_matrix, matrix_to_rpy, xyz_rpy_to_matrix, matrix_to_xyz_rpy)
 from .version import __version__
 
 __all__ = [
