+++ urchin/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.0.27"
+__version__ = "0.0.30"
